+++ test/pushgatewayTest.js
@@ -79,6 +79,23 @@ describe.each([
 					}),
 				).rejects.toThrow('push failed with status 400');
 			});
+
+			it('should timeout when taking too long', () => {
+				const mockHttp = nock('http://192.168.99.100:9091')
+					.post('/metrics/job/testJob/key/va%26lue', body)
+					.delay(100)
+					.reply(200);
+
+				expect.assertions(1);
+				return instance
+					.pushAdd({
+						jobName: 'testJob',
+						groupings: { key: 'va&lue' },
+					})
+					.catch(err => {
+						expect(err.message).toStrictEqual('Pushgateway request timed out');
+					});
+			});
 		});
 
 		describe('push', () => {
@@ -114,6 +131,18 @@ describe.each([
 					}),
 				).rejects.toThrow('push failed with status 400');
 			});
+
+			it('should timeout when taking too long', () => {
+				const mockHttp = nock('http://192.168.99.100:9091')
+					.put('/metrics/job/test%26Job', body)
+					.delay(100)
+					.reply(200);
+
+				expect.assertions(1);
+				return instance.push({ jobName: 'test&Job' }).catch(err => {
+					expect(err.message).toStrictEqual('Pushgateway request timed out');
+				});
+			});
 		});
 
 		describe('delete', () => {
@@ -135,6 +164,18 @@ describe.each([
 				return expect(instance.delete({ jobName: 'testJob' })).rejects.toThrow(
 					'push failed with status 400',
 				);
+			});
+
+			it('should timeout when taking too long', () => {
+				const mockHttp = nock('http://192.168.99.100:9091')
+					.delete('/metrics/job/testJob')
+					.delay(100)
+					.reply(200);
+
+				expect.assertions(1);
+				return instance.delete({ jobName: 'testJob' }).catch(err => {
+					expect(err.message).toStrictEqual('Pushgateway request timed out');
+				});
 			});
 		});
 
@@ -238,7 +279,7 @@ describe.each([
 
 		beforeEach(() => {
 			registry = undefined;
-			instance = new Pushgateway('http://192.168.99.100:9091');
+			instance = new Pushgateway('http://192.168.99.100:9091', { timeout: 10 });
 			const promClient = require('../index');
 			const cnt = new promClient.Counter({ name: 'test', help: 'test' });
 			cnt.inc(100);
@@ -254,7 +295,11 @@ describe.each([
 
 		beforeEach(() => {
 			registry = new Registry(regType);
-			instance = new Pushgateway('http://192.168.99.100:9091', null, registry);
+			instance = new Pushgateway(
+				'http://192.168.99.100:9091',
+				{ timeout: 10 },
+				registry,
+			);
 			const promeClient = require('../index');
 			const cnt = new promeClient.Counter({
 				name: 'test',